requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.5
django-cors-headers==3.13.0
django-elasticsearch-dsl==7.2.2
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
elasticsearch==7.17.8
elasticsearch-dsl==7.4.0
flower==1.2.0
humanize==4.4.0
iniconfig==1.1.1
kombu==5.2.4
packaging==22.0
Pillow==9.4.0
pluggy==1.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.7
redis==4.4.0
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
